+++ src/modules/components/proposal/index.ts
@@ -1,1 +1,2 @@
 export * from './proposalDataListItem';
+export * from './proposalActions';
