+++ src/modules/components/proposal/proposalActions/actions/index.ts
@@ -1,0 +1,1 @@
+export * from './proposalActionWithdrawToken';
