+++ src/modules/components/proposal/proposalActions/index.ts
@@ -1,1 +1,11 @@
-export * from './proposalActions';
+import { ProposalActionsContainer } from './proposalActionsContainer';
+import { ProposalActionsAction } from './proposalActionsAction';
+
+export const ProposalActions = {
+  Container: ProposalActionsContainer,
+  Action: ProposalActionsAction
+};
+
+export * from './actions';
+export * from './proposalActionsContext';
+export * from './proposalActionsTypes';
