+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/index.ts
