+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalAction.ts
